Offline_analizer/Platform/ToolInstaller/GUI_PC_py_requirements.txt
setuptools==75.1.0
psutil==6.0.0
pywin32==306
rpyc==6.0.0
colorama==0.4.6
numpy==2.1.2
pandas==2.2.2
regex==2024.5.15
openpyxl==3.1.5
Unidecode==1.3.8
PyYAML==6.0.1
simplejson==3.19.2
xlrd==2.0.1
easygui==0.98.3
fuzzywuzzy==0.18.0
dispatch==2.0.6
pysftp==0.2.9
fabric==3.2.2
Jinja2==3.1.4
GitPython==3.1.43
Python_CANoe==0.1
certifi==2024.7.4
gitdb==4.0.11
pytz==2024.2
six==1.16.0
smmap==5.0.1
tqdm==4.66.4
colorlog==6.8.2
logger==1.4
WMI==1.5.1
gitinfo==1.2.0
paramiko==3.4.0
bs4==0.0.2
matplotlib==3.9.2
pyserial==3.5
pygame==2.6.0
asammdf==7.4.5
requests==2.32.3
logidrivepy==0.2.0
lxml==5.2.2
pyinstaller==6.9.0
cffi==1.17.1
clr_loader==0.2.6
DateTime==5.5
importlib==1.0.4
pycparser==2.22
pythonnet==3.0.4
thread==2.0.5
tk==0.1.0
typing_extensions==4.12.2
zope.interface==7.0.3
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
kiwisolver==1.4.7
pillow==11.0.0
pyparsing==3.2.0
python_dateutil==2.9.0.post0
packaging==24.2
plotly==5.24.1
tenacity==9.0.0
elporto_rqm==1.0.0
python-can==4.5.0